requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
faiss-cpu==1.10.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
loguru==0.7.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
openai==1.70.0
packaging==24.2
pydantic==2.11.1
pydantic_core==2.33.0
pydot==3.0.4
pyparsing==3.2.3
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.1
tokenizers==0.21.1
# torch @ https://mirrors.aliyun.com/pytorch-wheels/cpu/torch-2.3.1+cpu-cp312-cp312-linux_x86_64.whl#sha256=2141a6cb7021adf2f92a0fd372cfeac524ba460bd39ce3a641d30a561e41f69a
torch@https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.6.0%2Bcpu.cxx11.abi-cp312-cp312-linux_x86_64.whl#sha256=a45917579548be5c7c97c04531161aa5e3b100e37b815bfdf54402fc0b4a4f36 ; sys_platform == 'linux'
torch==2.6.0 ; sys_platform == 'win32' or sys_platform == 'darwin'
tqdm==4.67.1
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
scipy==1.15.2 # network pagerank依赖
pyan3==1.2.0 # 解析 Python callgraph 使用